Mark libft2 as LLNDK
libft2 is used by libRS_internal.so which is a VNDK-SP library. Since
VNDK-SP lib can only depend on other VNDK-SP or LL-NDK libraries, this
has to be marked as either of them.

Among the two choices, LLNDK is more appropriate since ...
1) freetype lib should be updated along with the platform when the new
font types are added or existing font types are modified.
2) The ABI of this lib is pretty stable.

Currently, only the small set of symbols that are actually needed by
libRS_internal.so is exposed via the LLNDK stub, so that we can remove
or alter symbols that are not exposed.

Bug: 38432260
Bug: 64425518
Test: BOARD_VNDK_VERSION=current m -j libRS_internal.vendor

Change-Id: I499b556e64ff2da9f6a86d5ebd8153bef6ad7090